Fix Accuracy Change Issue
