+++ factory-ai-vision/EdgeSolution/modules/WebModule/ui/src/store/project/projectActions.ts
@@ -1,4 +1,5 @@
 import Axios from 'axios';
+import * as R from 'ramda';
 import {
   ProjectThunk,
   GetProjectSuccessAction,
@@ -422,4 +423,36 @@ export const thunkUpdateAccuracyRange = (isDemo: boolean): ProjectThunk => (
     .catch((e) => {
       dispatch(postProjectFail(e, isDemo));
     });
+};
+
+export const thunkCheckAndSetAccuracyRange = (newSelectedParts: any[], isDemo: boolean) => (
+  dispatch,
+  getState,
+): void => {
+  const images = getState().images.filter((e) => !e.is_relabel);
+
+  const partsWithImageLength = images.reduce((acc, cur) => {
+    const { id } = cur.part;
+    const relatedPartIdx = acc.findIndex((e) => e.id === id);
+    if (relatedPartIdx >= 0) acc[relatedPartIdx].length = acc[relatedPartIdx].length + 1 || 1;
+    return acc;
+  }, R.clone(newSelectedParts));
+
+  const minimumLengthPart = partsWithImageLength.reduce(
+    (acc, cur) => {
+      if (cur.length < acc.length) return { name: cur.name, length: cur.length };
+      return acc;
+    },
+    { name: '', length: Infinity },
+  );
+  if (minimumLengthPart.length === Infinity) return;
+  if (minimumLengthPart.length < 30) {
+    dispatch(updateProjectData({ accuracyRangeMax: 40, accuracyRangeMin: 10 }, isDemo));
+  } else if (minimumLengthPart.length >= 30 && minimumLengthPart.length < 80) {
+    dispatch(updateProjectData({ accuracyRangeMax: 60, accuracyRangeMin: 30 }, isDemo));
+  } else if (minimumLengthPart.length >= 80 && minimumLengthPart.length < 130) {
+    dispatch(updateProjectData({ accuracyRangeMax: 80, accuracyRangeMin: 50 }, isDemo));
+  } else if (minimumLengthPart.length >= 130) {
+    dispatch(updateProjectData({ accuracyRangeMax: 90, accuracyRangeMin: 60 }, isDemo));
+  }
 };
